test: reformat testing structure
